pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "blog-narrator"
version = "0.1.0"
description = "Add AI-narrated audio to your blog posts. Local TTS via Kokoro-82M on Apple Silicon."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Gal Sapir" }]
keywords = ["tts", "jekyll", "blog", "narration", "kokoro", "mlx"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
    "mlx-audio>=0.3",
    "mutagen>=1.47",
    "pyyaml>=6.0",
]

[project.scripts]
narrate = "blog_narrator.cli:main"

[project.urls]
Homepage = "https://github.com/galsapir/blog-narrator"
Repository = "https://github.com/galsapir/blog-narrator"
